오픈소스 테스트 코드 분리 [#4]
