+++ src/test/kotlin/com/osscube/api/application/controller/open_source/SaveOpenSourceTest.kt
@@ -1,4 +1,4 @@
-package com.osscube.api.application.controller
+package com.osscube.api.application.controller.open_source
 
 import com.fasterxml.jackson.databind.ObjectMapper
 import com.osscube.api.application.request.OpenSourceSaveRequest
@@ -6,8 +6,7 @@ import com.osscube.api.config.TestContainers
 import com.osscube.api.domain.dto.OpenSourceSaveRequestDto
 import com.osscube.api.domain.model.entity.OpenSource
 import com.osscube.api.domain.model.repository.OpenSourceRepository
-import org.hamcrest.Matchers
-import org.hamcrest.Matchers.containsInRelativeOrder
+import org.hamcrest.Matchers.isA
 import org.junit.jupiter.api.AfterEach
 import org.junit.jupiter.api.DisplayName
 import org.junit.jupiter.api.Test
@@ -17,13 +16,13 @@ import org.springframework.boot.test.context.SpringBootTest
 import org.springframework.http.HttpStatus
 import org.springframework.http.MediaType
 import org.springframework.test.web.servlet.MockMvc
-import org.springframework.test.web.servlet.request.MockMvcRequestBuilders
+import org.springframework.test.web.servlet.request.MockMvcRequestBuilders.post
 import org.springframework.test.web.servlet.result.MockMvcResultMatchers.jsonPath
 import org.springframework.test.web.servlet.result.MockMvcResultMatchers.status
 
 @SpringBootTest
 @AutoConfigureMockMvc
-class OpenSourceControllerTest : TestContainers() {
+class SaveOpenSourceTest : TestContainers() {
     @Autowired
     private lateinit var mockMvc: MockMvc
 
@@ -47,13 +46,12 @@ class OpenSourceControllerTest : TestContainers() {
         // when // then
         mockMvc
             .perform(
-                MockMvcRequestBuilders
-                    .post("/api/v1/open-sources")
+                post("/api/v1/open-sources")
                     .contentType(MediaType.APPLICATION_JSON)
                     .content(objectMapper.writeValueAsString(request))
             )
             .andExpect { status().isCreated }
-            .andExpect(jsonPath("$.openSource.openSourceId").value(Matchers.isA(String::class.java), String::class.java))
+            .andExpect(jsonPath("$.openSource.openSourceId").value(isA(String::class.java), String::class.java))
             .andExpect(jsonPath("$.openSource.name").value("name"))
             .andExpect(jsonPath("$.openSource.originUrl").value("origin url"))
     }
@@ -69,8 +67,7 @@ class OpenSourceControllerTest : TestContainers() {
         // when // then
         mockMvc
             .perform(
-                MockMvcRequestBuilders
-                    .post("/api/v1/open-sources")
+                post("/api/v1/open-sources")
                     .contentType(MediaType.APPLICATION_JSON)
                     .content(objectMapper.writeValueAsString(request))
             )
@@ -78,29 +75,5 @@ class OpenSourceControllerTest : TestContainers() {
             .andExpect(jsonPath("$.status").value(HttpStatus.CONFLICT.value()))
             .andExpect(jsonPath("$.error").value("OPEN_SOURCE-002"))
             .andExpect(jsonPath("$.message").value("이미 저장된 오픈소스입니다."))
-    }
-
-    @DisplayName("오픈소스 목록을 조회한다.")
-    @Test
-    fun getOpenSources() {
-        // given
-        val given = listOf(
-            OpenSource("name1", "origin url"),
-            OpenSource("name2", "origin url"),
-            OpenSource("name3", "origin url")
-        )
-        openSourceRepository.saveAll(given)
-
-        // when // then
-        mockMvc
-            .perform(
-                MockMvcRequestBuilders
-                    .get("/api/v1/open-sources")
-            )
-            .andExpect(status().isOk)
-            .andExpect(jsonPath("$.openSources").isArray)
-            .andExpect(jsonPath("$.openSources.length()").value(given.size))
-            .andExpect(jsonPath("$.openSources[*].name").value(containsInRelativeOrder("name1", "name2", "name3")))
-            .andExpect(jsonPath("$.openSources[*].originUrl").value(containsInRelativeOrder("origin url", "origin url", "origin url")))
     }
 }
